docs: modify twitter link Footer.vue
The old twitter link, "https://twitter.com/startalehq" is no longer active. Instead, modified the link to https://x.com/StartaleGroup".

Meanwhile, Twitter has rebranded as X. It's appropriate to call it X now, and the icon needs to change accordingly. It looks like this needs to be fixed as well.